requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-And-Redis==1.0.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
redis==3.5.3
toml==0.10.2
Werkzeug==2.0.1
arrow~=1.1.1
